for v1.0.9
・New function 'Correlation analysis' in Normalized count analysis.
・Enhance usability
1. 'Select samples' function in Pair-wise DEG, 3 conditions DEG, and Normalized count analysis.
You can choose the samples you want to analyze using the 'Select samples' function in Pair-wise DEG and 3 conditions DEG, and Normalized count analysis.
It's not necessary to create a count file containing only the samples you intend to analyze, except for the batch-mode in Pair-wise DEG.
2. k-means clustering in Multi DEG and Normalized count analysis.
When genes of interest are selected in the 'kmeans_result (Multi DEG)' and 'k-means clustering result (Normalized count analysis)' panel, their positions are displayed on the heatmap.
Moreover, the order of clusters on the heatmap can be changed using the 'Order of clusters on the heatmap' function.
 img(src="norm k-means 2.png", width = 600,height = 400)
3. 'Order of groups' in the enrichment analysis.
In the previous versions, the x-axis in the dotplot was arranged alphabetically. you have the flexibility to change the order of groups.
4. Add MA plot in GOI profiling of Pair-wise DEG.
5. Rasterize Heatmap and volcano plot to reduce data size.
・Improve the reproducibility of k-means clustering. typically relies on the initial cluster center, which poses a reproducibility issue.
In the previous versions, to mitigate this problem, k-means clustering was executed 100 times, and a final consensus k-means clustering was employed. However, this approach could not gurantee complete reproducibility.
In the current version, to adress this problem, k-means clustering is performed with a fixed initial cluster center.
・Fix bug
1. bug regarding input format 'normalized count data + meta data' in Normalized count analysis.
2. bug when using ENSEMBL ID as gene names.
    - Some table data formatting was skewed.
    - 'Option: Select a normalized count file' in Pair-wise DEG and 3 conditions DEG was not work.